+++ app/src/main/java/com/example/cryptoapp/domain/CoinRepository.kt
@@ -5,5 +5,5 @@ import androidx.lifecycle.LiveData
 interface CoinRepository {
     fun getCoinInfoList(): LiveData<List<CoinInfo>>
     fun getCoinInfo(fromSymbol: String): LiveData<CoinInfo>
-    suspend fun loadData()
+    fun loadData()
 }
