+++ app/src/main/java/com/example/cryptoapp/domain/LoadDataUseCase.kt
@@ -3,5 +3,5 @@ package com.example.cryptoapp.domain
 class LoadDataUseCase(
     private val repository: CoinRepository
 ) {
-    suspend operator fun invoke() = repository.loadData()
+    operator fun invoke() = repository.loadData()
 }
